Move theme to lib
