+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@temporalio/ui",
-  "version": "2.23.2",
+  "version": "2.23.3-theme.0",
   "engines": {
     "pnpm": ">=8.6.0",
     "node": ">=18.15.0"
